Got rid of a warning on Linux x86_64
git-svn-id: https://hypertable.googlecode.com/svn/trunk@484 01261837-d133-0410-8dcb-6792eb736156

+++ src/cc/Hypertable/RangeServer/AccessGroup.cc
@@ -113,7 +113,7 @@ void AccessGroup::get_cached_rows(std::vector<std::string> &rows) {
 
 uint64_t AccessGroup::disk_usage() {
   boost::mutex::scoped_lock lock(m_mutex);
-  return m_disk_usage + (m_compression_ratio * (float)m_cell_cache_ptr->memory_used());
+  return m_disk_usage + (uint64_t)(m_compression_ratio * (float)m_cell_cache_ptr->memory_used());
 }
 
 bool AccessGroup::needs_compaction() {
